requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.13
Django==5.0.14
djangorestframework==3.16.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
langcodes==3.5.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
nltk==3.8.1
numpy==1.26.4
packaging==25.0
pandas==2.1.4
preshed==3.0.12
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
scikit-learn==1.4.2
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.5
srsly==2.5.2
thinc==8.3.10
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.20.1
typer-slim==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
